requirements.txt
numpy
scipy
fastdtw
matplotlib
imageio
google-cloud-storage
flask
opencv-python-headless
mediapipe
pandas